requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.2.2
enum34==1.1.6
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
pycparser==2.18
pyOpenSSL==18.0.0
six==1.11.0
Twisted==18.4.0
zope.interface==4.5.0
